Some more cleanup

+++ src/RheelEngine/Animator/Transition.cpp
@@ -11,7 +11,7 @@ static float transitionQuadraticOut(float t) { return 1.0f - (t - 1.0f) * (t - 1
 static float transitionCubic(float t) { return -2.0f * t * t * t + 3.0f * t * t; }
 static float transitionCubicIn(float t) { return t * t * t; }
 static float transitionCubicOut(float t) { return t * t * t - 3 * t * t + 3 * t; }
-static float transitionCosine(float t) { return (1.0f - std::cos(M_PI * t)) / 2.0f; }
+static float transitionCosine(float t) { return (1.0f - static_cast<float>(std::cos(M_PI * t))) / 2.0f; }
 
 const Transition Transition::LINEAR(transitionLinear);
 const Transition Transition::QUADRATIC_IN(transitionQuadraticIn);
